requirements.txt
asgiref==3.8.1
certifi==2024.8.30
channels==4.1.0
charset-normalizer==3.4.0
Django==5.1.2
django-cors-headers==4.6.0
djangorestframework==3.15.2
drf-yasg==1.21.8
geographiclib==2.0
geopy==2.4.1
idna==3.10
inflection==0.5.1
mysqlclient==2.2.5
packaging==24.1
pyproj==3.7.0
python-decouple==3.8
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
sqlparse==0.5.1
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
